move processing counters to consensus core and remove rpc service crate dep on consensus
